fixed most changes except dual threading still todo
