+++ samples/ChangefeedMigrationSample/ChangefeedMigrationSample.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{3ACC802F-8C9D-43C2-B945-F1F48F990586}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Changefeedconsole</RootNamespace>
-    <AssemblyName>Changefeedconsole</AssemblyName>
+    <RootNamespace>ChangeFeedMigrationSample</RootNamespace>
+    <AssemblyName>ChangeFeedMigrationSample</AssemblyName>
     <TargetFrameworkVersion>v4.7</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -69,6 +69,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DeviceReading.cs" />
+    <Compile Include="DocumentFeedObserver.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
