+++ samples/ChangefeedMigrationSample/Readme.md.txt
@@ -1,0 +1,4 @@
+Code sample 
+Simple migration tool using change feed processor library. 
+
+.NET 4.7
