--- samples/ChangefeedMigrationSample/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Azure.DocumentDB" version="1.14.1" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net47" />
-</packages>
